+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/java/Java.java
@@ -80,6 +80,10 @@ public class Java extends FormatterFactory {
 		addStepFactory(removeWildcardImports);
 	}
 
+	public void addReplaceObsoletes(ReplaceObsoletes replaceObsoletes) {
+		addStepFactory(replaceObsoletes);
+	}
+
 	public void addFormatAnnotations(FormatAnnotations formatAnnotations) {
 		addStepFactory(formatAnnotations);
 	}